Unfortunately, it turns out that JavaScript overlay types have a few more restrictions than I initially thought. For example, JavaScriptObject types cannot implement interfaces.
Getting around this limitation required creating an Interface, a JSO implementation, and a wrapper implementation (of the interface) for the JSO. This turned out to be quite an extensive amount of code which only served to bring full Java semantics to a JavaScript object that would, in turn, be compiled back to JavaScript.

In light of that realization (and because the types used in this project are not likely to ever be used with a server-side implementation), I am fine implementing them as JSO-dependent classes.